forgor that

+++ src/events/messageCreate.ts
@@ -86,9 +86,8 @@ export default async (client: CustomClient, message: Message): Promise<void> => 
 
   if (
     !canSendLinks &&
-    hasLink.hasUrls
-    // &&
-    // level < 3 // Moderators can post links anywhere
+    hasLink.hasUrls &&
+    level < 3 // Moderators can post links anywhere
   ) {
     await message.delete().catch(() => {});
     return;
